refactor: Rename `send_command.rs` to `server_send.rs`
